packages/core: rename AuthHelper to AuthConnector
